支持微积分计算.

+++ jstarcraft-ai-math/src/main/java/com/jstarcraft/ai/math/MathUtility.java
@@ -22,7 +22,7 @@ public class MathUtility {
      * @param base
      * @return
      */
-    public static float logarithm(float number, int base) {
+    public static float logarithm(float number, float base) {
         return (float) (Math.log(number) / Math.log(base));
     }
